Moonjump/Devil Doom fixes
Recoded Moonjump, now it properly works off an inclusive L press instead of exclusive, and is no longer a C2 code.
The Jump Dash physics changes were all consolidated to one code that checks if the player is in stg0710, if so then all Jump Dash physics are changed back to default.
And lastly, did more cleanup of the ini, and got rid of the WIP section.