Betray: Beetle prompt
Edit to all the lines of Hero partners telling Shadow to attack Beetles with his homing attack, in cases where they would scold you immediately for doing so.
In all cases, their lines were changed to Shadow's version.
Stages changed;
100: Sonic
200: Rouge
202: Sonic
302: Tails
404: Rouge
501: Tails
502: Omega
Maria is an exception because she doesn't scold you in The Doom, where it's somewhat necessary depending on the route.