Westopolis red rings v2
Moved the vanilla ring position in the tiny alleyway to helicopter section. Added a subtle ramp in the ground for that section.

The old alleyway ring position now has an additional target switch in its place for the caged red ring.

Added the "... we'll get the last laugh!" line, should play after old key door position.